requirements.txt
Flask==2.2.3
flask-sqlalchemy==3.0.3
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-dotenv==1.0.0
apscheduler==3.10.1
pytz==2023.3
flask-socketio==5.3.3
simple-websocket==0.10.1
eventlet==0.33.3
python-engineio==4.4.1
python-socketio==5.8.0
werkzeug==2.2.3
boto3==1.26.135
requests==2.29.0 